requirements.txt
# Core dependencies
python-telegram-bot>=20.0,<21.0
PyYAML>=6.0
psutil>=5.9.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
